pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "frontier-interp"
version = "0.1.0"
description = "Config-driven experiments for mechanistic compressibility and self-interpretability gaps"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "torch",
  "transformers",
  "datasets",
  "pyyaml",
  "numpy",
  "pandas",
  "matplotlib",
  "scipy",
  "tqdm",
  "accelerate",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
